requirements.txt
Django
djangorestframework
celery
django-celery-beat
channels
djangorestframework-simplejwt
django-redis
Requests
ccxt